requirements.txt
streamlit
ultralytics
streamlit-webrtc
opencv-python-headless
av
Pillow
numpy
